Trying to fix tests not executing setUpModule
